Add more files - several 64bit portability issues in this codebase

+++ daphne_on_Android/Daphne_on_Android/Daphne_on_Android.NativeActivity/daphne-1.0-src/ldp-in/ldv1000.cpp
@@ -610,9 +610,9 @@ unsigned int g_ldv1000_last_event = 0;
 
 void ldv1000_event_callback(void *eventType)
 {
-	g_ldv1000_last_event = (unsigned long) eventType;	// changed to long for x64 support (thanks Joker)
+	g_ldv1000_last_event = (uintptr_t) eventType;	// changed to long for x64 support (thanks Joker)
 
-	switch ((unsigned long) eventType)
+	switch ((uintptr_t) eventType)
 	{
 	case LDV1000_EVENT_VSYNC_END:
 #ifdef DEBUG
